+++ minecraft/net/minecraft/src/universalelectricity/components/TileEntityElectricFurnace.java
@@ -12,7 +12,7 @@ public class TileEntityElectricFurnace extends TileEntity implements ITexturePro
 	public static final int smeltingTimeRequired = 160;
 		
 	//The electricity stored in this tile entity
-	public int electricityStored = 0;
+	public double electricityStored = 0.0;
 	
 	//How many ticks has this item been smelting for?
 	public int smeltingTicks = 0;
@@ -41,7 +41,7 @@ public class TileEntityElectricFurnace extends TileEntity implements ITexturePro
 	 * @return watts - The amount of rejected power to be sent back into the conductor
 	 */
     @Override
-    public int onReceiveElectricity(int watts, int voltage, byte side)
+    public double onReceiveElectricity(double watts, int voltage, byte side)
     {
     	if(voltage > this.getVolts())
     	{
@@ -51,8 +51,8 @@ public class TileEntityElectricFurnace extends TileEntity implements ITexturePro
     	//Only accept electricity from the front side
     	if(canReceiveElectricity(side) || side == -1)
 		{
-	    	int rejectedElectricity = Math.max((this.electricityStored + watts) - this.getElectricityCapacity(), 0);
-			this.electricityStored = Math.max(this.electricityStored+watts - rejectedElectricity, 0);
+    		double rejectedElectricity = Math.max((this.electricityStored + watts) - this.getElectricityCapacity(), 0.0);
+			this.electricityStored = Math.max(this.electricityStored+watts - rejectedElectricity, 0.0);
 			return rejectedElectricity;
 		}
     	return watts;
@@ -89,7 +89,7 @@ public class TileEntityElectricFurnace extends TileEntity implements ITexturePro
 			           	
 		            	if(electricItem.canProduceElectricity())
 			           	{
-			            	int receivedElectricity = electricItem.onUseElectricity(electricItem.getTransferRate(), this.containingItems[0]);
+			            	double receivedElectricity = electricItem.onUseElectricity(electricItem.getTransferRate(), this.containingItems[0]);
 			            	this.onReceiveElectricity(receivedElectricity, electricItem.getVolts(), (byte)-1);
 			            }
 		            }
@@ -167,7 +167,7 @@ public class TileEntityElectricFurnace extends TileEntity implements ITexturePro
 	public void readFromNBT(NBTTagCompound par1NBTTagCompound)
     {
     	super.readFromNBT(par1NBTTagCompound);
-    	this.electricityStored = par1NBTTagCompound.getInteger("electricityStored");
+    	this.electricityStored = par1NBTTagCompound.getDouble("electricityStored");
     	this.facingDirection = par1NBTTagCompound.getByte("facingDirection");
     	this.smeltingTicks = par1NBTTagCompound.getInteger("smeltingTicks");
     	NBTTagList var2 = par1NBTTagCompound.getTagList("Items");
@@ -189,7 +189,7 @@ public class TileEntityElectricFurnace extends TileEntity implements ITexturePro
 	public void writeToNBT(NBTTagCompound par1NBTTagCompound)
     {
     	super.writeToNBT(par1NBTTagCompound);
-    	par1NBTTagCompound.setInteger("electricityStored", this.electricityStored);
+    	par1NBTTagCompound.setDouble("electricityStored", this.electricityStored);
     	par1NBTTagCompound.setByte("facingDirection", this.facingDirection);
     	par1NBTTagCompound.setInteger("smeltingTicks", this.smeltingTicks);
     	NBTTagList var2 = new NBTTagList();
@@ -209,14 +209,14 @@ public class TileEntityElectricFurnace extends TileEntity implements ITexturePro
 	 * @return Return the stored electricity in this consumer. Called by conductors to spread electricity to this unit.
 	 */
     @Override
-	public int getStoredElectricity()
+	public double getStoredElectricity()
     {
     	return this.electricityStored;
     }
     @Override
-    public int getElectricityCapacity()
+    public double getElectricityCapacity()
 	{
-		return 1800;
+		return 1800.0;
 	}
 	@Override
 	public int getStartInventorySide(int side)
@@ -352,7 +352,7 @@ public class TileEntityElectricFurnace extends TileEntity implements ITexturePro
         {
         	int packetID = dataStream.readInt();
         	this.facingDirection = (byte)dataStream.readDouble();
-        	this.electricityStored = (int)dataStream.readDouble();
+        	this.electricityStored = dataStream.readDouble();
         	this.smeltingTicks = (int)dataStream.readDouble();
         	this.disableTicks = (int)dataStream.readDouble();
         }
